Advanced_apps/Advanced_Strategic_Finance_Agent/requirements.txt
langchain_community==0.4.1
langchain_core==1.2.21
langchain_google_genai==4.2.1
langchain_tavily==0.2.17
langgraph==1.1.3
pytest==9.0.2
python-dotenv==1.2.2
Requests==2.32.5
streamlit==1.46.0
